New Xelthia Surgical Jacket + Surgery Check for Xelthia Jackets (#2340)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Title. Allows the insulated jacket to be used as gloves during surgery,
adding a component to facilitate this, and adds a similar jacket for the
express purpose of being used in surgery.

Also comes with a bugfix for the Xelthia-specific insulated jacket
sprite not working properly and a bit of fine tuning for its
description. The jackets are also added to appropriate spots like CE and
CMO lockers.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

A fair number of screenshots taken while testing this at various points

![image](https://github.com/user-attachments/assets/1e903668-e109-4c7c-b136-9ff3b16f277a)

![image](https://github.com/user-attachments/assets/2793b4d0-21f3-41bd-8934-aae79f40b73a)
(pre-fix for xelthia insulated jacket sprite)

![image](https://github.com/user-attachments/assets/87f6c976-a32a-43f6-ada0-c6ca6549841d)

![image](https://github.com/user-attachments/assets/f314e9bc-cc85-407f-9b67-e1f6b79313d6)

![image](https://github.com/user-attachments/assets/47d2dc8a-96d7-46df-b222-eabf02bc6306)

</p>
</details>

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Zeta_Null
- add: Added a surgical jacket, for Xelthia to perform surgery wearing
in place of gloves.
- tweak: Tweaked surgery to allow for jackets with gloves to be treated
as gloves.